added cloudwatch logging

+++ hdfs_to_s3/local_to_s3_luigi.py
@@ -9,18 +9,41 @@ import uuid
 import os
 from logging.handlers import RotatingFileHandler
 from os import walk
+import watchtower
+from boto3.session import Session
 
-log = logging.getLogger("luigi-interface")
-job_uuid = hex(uuid.uuid1().int >> 96)[2:]
-formatter = logging.Formatter('<LOGMSGSTART>' +
-                              '%(asctime)s - ' + job_uuid +
-                              ' - %(levelname)s - %(filename)s - %(lineno)s - %(message)s')
+
 pid = os.getpid()
+(aws_access_key_id,
+ aws_secret_access_key,
+ aws_session_token,
+ session_token_expiry) = get_aws_credentials()
+
+boto3_session = Session(aws_access_key_id=aws_access_key_id,
+                        aws_secret_access_key=aws_secret_access_key,
+                        aws_session_token=aws_session_token,
+                        region_name='us-west-2'
+                        )
+
+log = logging.getLogger(str(pid))
+formatter = logging.Formatter(f'{pid} '
+                              f'%(asctime)s - '
+                              f' - %(levelname)s - %(filename)s - %(lineno)s - %(message)s')
+
 dirname = os.path.dirname(__file__)
-filename = os.path.join(dirname, f'logs/{pid}_local_to_s3.log')
-fh = RotatingFileHandler(filename, maxBytes=1000000, backupCount=10)
+filename = os.path.join(dirname, f'logs/local_to_s3_{pid}.log')
+fh = RotatingFileHandler(filename, maxBytes=1000000, backupCount=1)
 fh.setFormatter(formatter)
 log.addHandler(fh)
+
+handler = watchtower.CloudWatchLogHandler(boto3_session=boto3_session,
+                                          log_group='shopping_midt_s3',
+                                          stream_name=f'shopping_midt_s3_{pid}',
+                                          create_log_group=True,
+                                          create_log_stream=True)
+handler.setFormatter(formatter)
+log.addHandler(handler)
+log.setLevel(logging.INFO)
 
 
 def log_timer(func):
@@ -37,6 +60,7 @@ def log_timer(func):
 
 
 class TaskTemplate(luigi.Task):
+    job_uuid = luigi.Parameter()
     env_config = luigi.DictParameter(visibility=luigi.parameter.ParameterVisibility.HIDDEN)
     year = luigi.Parameter()
     month = luigi.Parameter()
@@ -62,7 +86,7 @@ class TaskTemplate(luigi.Task):
 
     def get_token_path(self):
         token_path = (self.env_config['s3_luigi_status_location']
-                      + self.task_family + '/'
+                      + self.get_task_family() + '/'
                       + self.year + '_'
                       + self.month + '_'
                       + self.day + '_'
@@ -73,7 +97,7 @@ class TaskTemplate(luigi.Task):
     def complete(self):
         if not self.iscomplete:
             token_path = self.get_token_path()
-            log.info('checking to see if task is complete ' + token_path)
+            log.info(f'{self.job_uuid} checking to see if task is complete ' + token_path)
 
             if self.get_s3fs().exists(token_path):
                 log.info("it's complete")
@@ -92,7 +116,8 @@ class TaskTemplate(luigi.Task):
         self.iscomplete = True
 
 
-class DownloadOneFileFromHDFSRequest(luigi.ExternalTask):
+class DownloadOneFileFromHDFSResponse(luigi.ExternalTask):
+    job_uuid = luigi.Parameter()
     env_config = luigi.DictParameter(visibility=luigi.parameter.ParameterVisibility.HIDDEN)
     year = luigi.Parameter()
     month = luigi.Parameter()
@@ -123,6 +148,8 @@ class DownloadOneFileFromHDFSRequest(luigi.ExternalTask):
         return token_path
 
     def complete(self):
+        log.info(f'DownloadOneFileFromHDFSResponse.complete: job_uuid {self.job_uuid} pid {os.getpid()}')
+
         token_path = self.get_token_path()
         log.info('checking to see if task is complete ' + token_path)
 
@@ -134,8 +161,8 @@ class DownloadOneFileFromHDFSRequest(luigi.ExternalTask):
             return False
 
 
-class UploadOneFiletoS3Request(TaskTemplate):
-    resources = {'s3': 1}
+class UploadOneFiletoS3Response(TaskTemplate):
+    # resources = {'s3': 1}
 
     def get_s3_file_path(self):
         s3_file_path = self.env_config['s3_response_file_path'].format(self.year,
@@ -145,24 +172,29 @@ class UploadOneFiletoS3Request(TaskTemplate):
                                                                        self.filename)
         return s3_file_path
 
-    # def requires(self):
-    #     s3_file_path = self.get_s3_file_path()
-    #     log.info(f'checking for existing file on s3: {s3_file_path}')
-    #     s3 = self.get_s3fs()
-    #     if not s3.exists(s3_file_path):
-    #         return DownloadOneFileFromHDFSRequest(env_config=self.env_config,
-    #                                               year=self.year,
-    #                                               month=self.month,
-    #                                               day=self.day,
-    #                                               hour=self.hour,
-    #                                               filename=self.filename
-    #                                               )
-    #     else:
-    #         log.warning(f'file already exists on s3 - skipping download for {s3_file_path}')
-    #         return None
+    def requires(self):
+        log.info(f'UploadOneFiletoS3Response.requires: job_uuid {self.job_uuid} pid {os.getpid()}')
+
+        s3_file_path = self.get_s3_file_path()
+        log.info(f'checking for existing file on s3: {s3_file_path}')
+        s3 = self.get_s3fs()
+        if not s3.exists(s3_file_path):
+            return DownloadOneFileFromHDFSResponse(
+                job_uuid=self.job_uuid,
+                env_config=self.env_config,
+                year=self.year,
+                month=self.month,
+                day=self.day,
+                hour=self.hour,
+                filename=self.filename
+            )
+        else:
+            log.warning(f'file already exists on s3 - skipping download for {s3_file_path}')
+            return None
 
     @log_timer
     def run(self):
+        log.info(f'UploadOneFiletoS3Response.run: job_uuid {self.job_uuid} pid {os.getpid()}')
         local_directory = 'data'
         s3_file_path = self.get_s3_file_path()
         log.info(f'uploading file to s3: {s3_file_path}')
@@ -174,33 +206,48 @@ class UploadOneFiletoS3Request(TaskTemplate):
         self.mark_complete()
 
 
-class DeleteLocalFileRequest(TaskTemplate):
+class DeleteLocalFileResponse(TaskTemplate):
 
     def requires(self):
-        return UploadOneFiletoS3Request(env_config=self.env_config,
-                                        year=self.year,
-                                        month=self.month,
-                                        day=self.day,
-                                        hour=self.hour,
-                                        filename=self.filename,
-                                        )
+        log.info(f'DeleteLocalFileResponse.requires: job_uuid {self.job_uuid} pid {os.getpid()}')
+        return UploadOneFiletoS3Response(
+            job_uuid=self.job_uuid,
+            env_config=self.env_config,
+            year=self.year,
+            month=self.month,
+            day=self.day,
+            hour=self.hour,
+            filename=self.filename,
+        )
+
+    def complete(self):
+        return self.iscomplete
 
     def run(self):
-        local_file_path = os.path.join('data', self.filename)
-        log.info('deleting local file: ' + local_file_path)
-        try:
-            os.remove(local_file_path)
-        except FileNotFoundError as e:
-            log.warning(e)
-        self.mark_complete()
+        log.info(f'DeleteLocalFileResponse.run: job_uuid {self.job_uuid} pid {os.getpid()}')
+        local_file_names = []
+        for (dirpath, dirnames, filenames) in walk('data'):
+            local_file_names.extend(filenames)
+
+        for file_name in local_file_names:
+            if self.filename in file_name:
+                local_file_path = os.path.join('data', self.filename)
+                log.info('deleting local file: ' + local_file_path)
+                try:
+                    os.remove(local_file_path)
+                except FileNotFoundError as e:
+                    log.warning(e)
+        self.iscomplete = True
 
 
 class Find_All_Local_Files_For_Hour(TaskTemplate):
 
     def complete(self):
+        log.info(f'Find_All_Local_Files_For_Hour.complete: job_uuid {self.job_uuid} pid {os.getpid()}')
         return self.iscomplete
 
     def requires(self):
+        log.info(f'Find_All_Local_Files_For_Hour.requires: job_uuid {self.job_uuid} pid {os.getpid()}')
         log.info(f'looking for local files {self.year} {self.month} {self.day} {self.hour}')
 
         local_file_names = []
@@ -212,13 +259,15 @@ class Find_All_Local_Files_For_Hour(TaskTemplate):
             if file_name != '.gitignore' and '.temp-' not in file_name:
                 log.info(f"creating task - {file_name}")
                 task_list.append(
-                    DeleteLocalFileRequest(env_config=self.env_config,
-                                           year=self.year,
-                                           month=self.month,
-                                           day=self.day,
-                                           hour=self.hour,
-                                           filename=file_name
-                                           )
+                    DeleteLocalFileResponse(
+                        job_uuid=self.job_uuid,
+                        env_config=self.env_config,
+                        year=self.year,
+                        month=self.month,
+                        day=self.day,
+                        hour=self.hour,
+                        filename=file_name
+                    )
                 )
 
         return task_list
@@ -229,7 +278,9 @@ class Find_All_Local_Files_For_Hour(TaskTemplate):
 
 if __name__ == '__main__':
     run_timestamp = str(datetime.datetime.now()).replace(' ', '_').replace(':', '_')
-    print(run_timestamp)
+    log.info(run_timestamp)
+    job_uuid = hex(uuid.uuid1().int >> 96)[2:]
+    log.info(f'creating new job_uuid {job_uuid}')
 
     luigi.build([
         Find_All_Local_Files_For_Hour(
@@ -238,10 +289,12 @@ if __name__ == '__main__':
             month='02',
             day='01',
             hour='23',
-            filename=''
+            filename='',
+            job_uuid=job_uuid
         )
     ],
         workers=2,
-        local_scheduler=True,
+        local_scheduler=False,
+        # scheduler_host='127.0.0.1',
         detailed_summary=False
     )
